Type define the root hls.js entrypoint.
Also configured the TS compiler to emit declaration files to `dist/declarations`.

Also, for a canary build this includes package.json `"types"` key to point at the declarations.
This should be removed before merging this branch as we do NOT want to ask regular users to use our typing data yet until it is more complete than DT.